+++ sdk/resourcemanager/datafactory/armdatafactory/CHANGELOG.md
@@ -1,5 +1,43 @@
 # Release History
 
+## 6.0.0 (2024-02-01)
+### Breaking Changes
+
+- Type of `AzureFunctionActivityTypeProperties.Headers` has been changed from `any` to `map[string]*string`
+- Type of `WebActivityTypeProperties.Headers` has been changed from `any` to `map[string]*string`
+- Type of `WebHookActivityTypeProperties.Headers` has been changed from `any` to `map[string]*string`
+- Field `ReadBehavior` of struct `SalesforceServiceCloudV2Source` has been removed
+- Field `ReadBehavior` of struct `SalesforceV2Source` has been removed
+
+### Features Added
+
+- New enum type `ExpressionV2Type` with values `ExpressionV2TypeBinary`, `ExpressionV2TypeConstant`, `ExpressionV2TypeField`, `ExpressionV2TypeUnary`
+- New enum type `ServiceNowV2AuthenticationType` with values `ServiceNowV2AuthenticationTypeBasic`, `ServiceNowV2AuthenticationTypeOAuth2`
+- New enum type `SnowflakeAuthenticationType` with values `SnowflakeAuthenticationTypeAADServicePrincipal`, `SnowflakeAuthenticationTypeBasic`, `SnowflakeAuthenticationTypeKeyPair`
+- New function `*ServiceNowV2LinkedService.GetLinkedService() *LinkedService`
+- New function `*ServiceNowV2ObjectDataset.GetDataset() *Dataset`
+- New function `*ServiceNowV2Source.GetCopySource() *CopySource`
+- New function `*ServiceNowV2Source.GetTabularSource() *TabularSource`
+- New function `*SnowflakeV2Dataset.GetDataset() *Dataset`
+- New function `*SnowflakeV2LinkedService.GetLinkedService() *LinkedService`
+- New function `*SnowflakeV2Sink.GetCopySink() *CopySink`
+- New function `*SnowflakeV2Source.GetCopySource() *CopySource`
+- New struct `ExpressionV2`
+- New struct `ServiceNowV2LinkedService`
+- New struct `ServiceNowV2LinkedServiceTypeProperties`
+- New struct `ServiceNowV2ObjectDataset`
+- New struct `ServiceNowV2Source`
+- New struct `SnowflakeLinkedV2ServiceTypeProperties`
+- New struct `SnowflakeV2Dataset`
+- New struct `SnowflakeV2LinkedService`
+- New struct `SnowflakeV2Sink`
+- New struct `SnowflakeV2Source`
+- New field `AuthenticationType` in struct `SalesforceServiceCloudV2LinkedServiceTypeProperties`
+- New field `IncludeDeletedObjects` in struct `SalesforceServiceCloudV2Source`
+- New field `AuthenticationType` in struct `SalesforceV2LinkedServiceTypeProperties`
+- New field `IncludeDeletedObjects` in struct `SalesforceV2Source`
+
+
 ## 5.0.0 (2024-01-26)
 ### Breaking Changes